requirements.txt
streamlit
pg8000
python-dotenv
pandas
plotly
python-dateutil
sqlalchemy